refactor: `BankType.toPresentation` 함수 이동 및 `formattedNumber`, `formattedExpired`를 확장 함수로 변경
