+++ app/src/androidTest/java/woowacourse/payments/screen/CardListScreenTest.kt
@@ -10,9 +10,10 @@ import org.junit.Test
 import org.junit.jupiter.api.assertAll
 import woowacourse.payments.CardUiModelFixture
 import woowacourse.payments.ui.model.CardUiModel
+import woowacourse.payments.ui.model.formattedExpired
+import woowacourse.payments.ui.model.formattedNumber
 import woowacourse.payments.ui.screen.cardList.CardListScreen
 import woowacourse.payments.ui.screen.cardList.CardListStateHolder
-import woowacourse.payments.ui.screen.cardList.CardListUiState
 
 class CardListScreenTest {
     @get:Rule
@@ -30,12 +31,7 @@ class CardListScreenTest {
     @Test
     fun 카드가_0개일_때_새로운_카드_등록_텍스트와_카드_추가_박스가_표시된다() {
         // given
-        val testStateHolder =
-            CardListStateHolder(
-                CardListUiState(
-                    cards = emptyList(),
-                ),
-            )
+        val testStateHolder = CardListStateHolder(initialCards = emptyList())
 
         // when
         composeTestRule.setContent {
@@ -56,11 +52,7 @@ class CardListScreenTest {
     fun 카드가_1개일_때_카드_정보와_카드_추가_박스가_표시된다() {
         // given
         val testStateHolder =
-            CardListStateHolder(
-                CardListUiState(
-                    cards = listOf(card1),
-                ),
-            )
+            CardListStateHolder(initialCards = listOf(card1))
 
         // when
         composeTestRule.setContent {
@@ -72,8 +64,8 @@ class CardListScreenTest {
 
         // then
         assertAll(
-            { composeTestRule.onNodeWithText(card1.formattedNumber).assertIsDisplayed() },
-            { composeTestRule.onNodeWithText(card1.formattedExpired).assertIsDisplayed() },
+            { composeTestRule.onNodeWithText(card1.formattedNumber()).assertIsDisplayed() },
+            { composeTestRule.onNodeWithText(card1.formattedExpired()).assertIsDisplayed() },
             { composeTestRule.onNodeWithText(card1.owner).assertIsDisplayed() },
             { composeTestRule.onNodeWithContentDescription("카드 추가").assertIsDisplayed() },
         )
@@ -83,12 +75,7 @@ class CardListScreenTest {
     fun 카드가_2개이상일_때_카드_정보와_TopBar에_추가_텍스트가_표시된다() {
         // given
         val cards = listOf(card1, card2)
-        val testStateHolder =
-            CardListStateHolder(
-                CardListUiState(
-                    cards = cards,
-                ),
-            )
+        val testStateHolder = CardListStateHolder(initialCards = cards)
 
         // when
         composeTestRule.setContent {
@@ -101,8 +88,8 @@ class CardListScreenTest {
         // then
         cards.forEach { card ->
             assertAll(
-                { composeTestRule.onNodeWithText(card.formattedNumber).assertIsDisplayed() },
-                { composeTestRule.onNodeWithText(card.formattedExpired).assertIsDisplayed() },
+                { composeTestRule.onNodeWithText(card.formattedNumber()).assertIsDisplayed() },
+                { composeTestRule.onNodeWithText(card.formattedExpired()).assertIsDisplayed() },
                 { composeTestRule.onNodeWithText(card.owner).assertIsDisplayed() },
             )
         }
